requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --no-annotate requirements.in
#
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
decorator==4.4.2
deepgram-sdk==2.3.0
dj-database-url==0.5.0
django==3.2.9
django-cors-headers==3.13.0
djangorestframework==3.12.4
frozenlist==1.3.3
idna==3.4
imageio==2.25.1
imageio-ffmpeg==0.4.8
moviepy==1.0.3
multidict==6.0.4
numpy==1.24.2
pedalboard==0.7.0
pillow==9.4.0
proglog==0.1.10
psycopg2-binary==2.9.2
pydub==0.25.1
pytz==2021.3
requests==2.28.2
sqlparse==0.4.2
tqdm==4.64.1
urllib3==1.26.14
websockets==10.4
yarl==1.8.2
